refactor: Extract shared BigSegmentStoreStatusProviderImpl into big_segments_common
Both big_segments.py and async_big_segments.py defined the class identically (bar one docstring word); move it to a shared module imported by both, matching the evaluator_common/client_common pattern.
